Mistakes are often the stepping stones to utter failure.

+++ gnome-base/librsvg/librsvg-2.40.21-r2.ebuild
@@ -6,7 +6,7 @@ VALA_USE_DEPEND="vapigen"
 inherit autotools gnome2 multilib-minimal vala
 
 DESCRIPTION="Scalable Vector Graphics (SVG) rendering library"
-HOMEPAGE="https://wiki.gnome.org/Projects/LibRsvg"
+HOMEPAGE="https://wiki.gnome.org/Projects/LibRsvg https://gitlab.gnome.org/GNOME/librsvg"
 
 LICENSE="LGPL-2+"
 SLOT="2"
@@ -94,7 +94,9 @@ multilib_src_install() {
 	gnome2_src_install
 
 	if ! use gtk-doc ; then
-		rm -r "${ED}"/usr/share/gtk-doc || die
+		if [[ -d "${ED}"/usr/share/gtk-doc ]] ; then
+			rm -r "${ED}"/usr/share/gtk-doc || die
+		fi
 	fi
 }
